+++ src/Modules/Workflows/Controllers/ScheduledActions.php
@@ -179,7 +179,7 @@ class ScheduledActions implements InitializableInterface
         $hook = $actionModel->getHook();
 
         switch ($hook) {
-            case WorkflowsHooksAbstract::ACTION_ASYNC_EXECUTE_NODE:
+            case WorkflowsHooksAbstract::ACTION_ASYNC_EXECUTE_STEP:
                 $step = $this->getStepFromActionId($row['ID']);
 
                 if (empty($step)) {
@@ -195,6 +195,7 @@ class ScheduledActions implements InitializableInterface
 
                 break;
 
+            case WorkflowsHooksAbstract::ACTION_UNSCHEDULE_RECURRING_STEP_ACTION:
             case WorkflowsHooksAbstract::ACTION_UNSCHEDULE_RECURRING_NODE_ACTION:
                 $title = __('Unschedule workflow recurring scheduled step', 'post-expirator');
                 break;
@@ -235,6 +236,7 @@ class ScheduledActions implements InitializableInterface
 
             switch ($hook) {
                 case WorkflowsHooksAbstract::ACTION_ASYNC_EXECUTE_NODE:
+                case WorkflowsHooksAbstract::ACTION_ASYNC_EXECUTE_STEP:
                     if (ScheduledActionModel::argsAreOnNewFormat((array) $args)) {
                         $scheduledStepModel = new WorkflowScheduledStepModel();
                         $scheduledStepModel->loadByActionId($actionId);
@@ -289,11 +291,11 @@ class ScheduledActions implements InitializableInterface
 
                     $next = $step['next'] ?? [];
 
-                    $nodeType = $this->stepTypesModel->getNodeType($step['node']['data']['name']);
+                    $stepType = $this->stepTypesModel->getStepType($step['node']['data']['name']);
 
                     $sourceHandles = [];
-                    if (! is_null($nodeType)) {
-                        $handlesSchema = $nodeType->getHandleSchema();
+                    if (! is_null($stepType)) {
+                        $handlesSchema = $stepType->getHandleSchema();
 
                         foreach ($handlesSchema['source'] as $handle) {
                             $sourceHandles[$handle['id']] = $handle['label'];
@@ -312,7 +314,7 @@ class ScheduledActions implements InitializableInterface
                             }
 
                             if (empty($stepLabel)) {
-                                $stepNodeType = $this->stepTypesModel->getNodeType($nextStep['node']['data']['name']);
+                                $stepNodeType = $this->stepTypesModel->getStepType($nextStep['node']['data']['name']);
                                 if (is_object($stepNodeType)) {
                                     $stepLabel = $stepNodeType->getLabel();
                                 }
@@ -361,6 +363,7 @@ class ScheduledActions implements InitializableInterface
                     $html = $argsText;
                     break;
 
+                case WorkflowsHooksAbstract::ACTION_UNSCHEDULE_RECURRING_STEP_ACTION:
                 case WorkflowsHooksAbstract::ACTION_UNSCHEDULE_RECURRING_NODE_ACTION:
                     $html = __('Workflow recurring scheduled action', 'post-expirator');
                     break;
@@ -558,7 +561,7 @@ class ScheduledActions implements InitializableInterface
         $actionModel = new ScheduledActionModel();
         $actionModel->loadByActionId($actionId);
 
-        if ($actionModel->getHook() !== WorkflowsHooksAbstract::ACTION_ASYNC_EXECUTE_NODE) {
+        if ($actionModel->getHook() !== WorkflowsHooksAbstract::ACTION_ASYNC_EXECUTE_STEP) {
             return;
         }