+++ src/Modules/Workflows/Domain/Steps/Actions/Runners/AddPostTermRunner.php
@@ -11,7 +11,7 @@ use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\AddPo
 class AddPostTermRunner implements StepRunnerInterface
 {
     /**
-     * @var NodeRunnerProcessorInterface
+     * @var StepProcessorInterface
      */
     private $stepProcessor;