+++ src/Modules/Workflows/Domain/Steps/Triggers/Runners/OnPostScheduleRunner.php
@@ -13,7 +13,7 @@ class OnPostScheduleRunner implements TriggerRunnerInterface
     use InfiniteLoopPreventer;
 
     /**
-     * @var NodeRunnerProcessorInterface
+     * @var StepProcessorInterface
      */
     private $stepProcessor;