+++ tests/Integration/Modules/Workflows/Domain/Engine/NodeRunners/Actions/CorePostDeactivateWorkflowTest.php
@@ -6,9 +6,8 @@ use PublishPress\Future\Core\HookableInterface;
 use PublishPress\Future\Framework\Logger\LoggerInterface;
 use PublishPress\Future\Modules\Workflows\Domain\Engine\NodeRunners\Actions\CorePostDeactivateWorkflow;
 use PublishPress\Future\Modules\Workflows\Domain\Engine\VariableResolvers\WorkflowResolver;
-use PublishPress\Future\Modules\Workflows\Interfaces\NodeRunnerProcessorInterface;
+use PublishPress\Future\Modules\Workflows\Interfaces\StepProcessorInterface;
 use PublishPress\Future\Modules\Workflows\Interfaces\RuntimeVariablesHandlerInterface;
-use PublishPress\Future\Modules\Workflows\Interfaces\WorkflowEngineInterface;
 use PublishPress\Future\Modules\Workflows\Models\PostModel;
 
 
@@ -68,7 +67,7 @@ class CorePostDeactivateWorkflowTest extends \lucatume\WPBrowser\TestCase\WPTest
         $workflows = $this->createWorkflows();
 
         $runner = new CorePostDeactivateWorkflow(
-            $this->makeEmpty(NodeRunnerProcessorInterface::class, [
+            $this->makeEmpty(StepProcessorInterface::class, [
                 'executeSafelyWithErrorHandling' => function ($step, $callback, ...$args) {
                     call_user_func($callback, $step, ...$args);
                 }