+++ CHANGELOG.md
@@ -3,7 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 4.3.1 - 30 Jan, 2025
+## [4.3.2] - 30 Jan, 2025
+
+### Fixed
+
+- Fix typo in the `WorkflowEngine` class.
+- Fix reference to deprecated classes and interfaces.
+- Fix translation of shortcode settings in the Display settings page.
+
+## [4.3.1] - 30 Jan, 2025
 
 ### Added
 
@@ -36,7 +44,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Step "Conditional" renamed to "Conditional Delay".
 - Changed the default step's slug to reflect the new step name and classes.
 - Changed the Schedule Delay step settings to be more intuitive.
-- Chagned some text in the workflow editor to be more user friendly.
+- Changed some text in the workflow editor to be more user friendly.
 
 ### Fixed